+++ examples/ios/objc/Migration/Examples/Example_v5.h
@@ -23,7 +23,7 @@
 
 #pragma mark - Schema
 
-int schemaVersion = 5;
+NSInteger schemaVersion = 5;
 
 // Changes from previous version:
 // - Change the `Address` from `Object` to `EmbeddedObject`.
@@ -39,7 +39,7 @@ typedef NS_ENUM(int, Kind) {
     cow
 };
 @property NSString *name;
-@property int kindValue;
+@property NSInteger kindValue;
 @property enum Kind kind;
 + (Pet* )petWithName:(NSString *)name kind:(enum Kind)kind;
 @end
@@ -69,7 +69,7 @@ RLM_ARRAY_TYPE(Pet)
 @class Address;
 @interface Person : RLMObject
 @property NSString *fullName;
-@property int age;
+@property NSInteger age;
 @property Address *address;
 @property RLMArray<Pet *><Pet> *pets;
 + (Person* )personWithFullName:(NSString *)fullName age:(int)age;
